# Re-use compiled regular expressions when using the validator