--- a/validators.go
+++ b/validators.go
@@ -12,6 +12,11 @@ import (
 	"strconv"
 )
 
+var (
+	upperAlphanumericRegex = regexp.MustCompile(`[^ A-Z0-9!"#$%&'()*+,-.\\/:;<>=?@\[\]^_{}|~]+`)
+	alphanumericRegex      = regexp.MustCompile(`[^ \w!"#$%&'()*+,-.\\/:;<>=?@\[\]^_{}|~]+`)
+)
+
 // validator is common validation and formating of golang types to ach type strings
 type validator struct{}
 
@@ -241,7 +246,7 @@ func (v *validator) isOriginatorStatusCode(code int) error {
 
 // isUpperAlphanumeric checks if string only contains ASCII alphanumeric upper case characters
 func (v *validator) isUpperAlphanumeric(s string) error {
-	if regexp.MustCompile(`[^ A-Z0-9!"#$%&'()*+,-.\\/:;<>=?@\[\]^_{}|~]+`).MatchString(s) {
+	if upperAlphanumericRegex.MatchString(s) {
 		return errors.New(msgUpperAlpha)
 	}
 	return nil
@@ -249,7 +254,7 @@ func (v *validator) isUpperAlphanumeric(s string) error {
 
 // isAlphanumeric checks if a string only contains ASCII alphanumeric characters
 func (v *validator) isAlphanumeric(s string) error {
-	if regexp.MustCompile(`[^ \w!"#$%&'()*+,-.\\/:;<>=?@\[\]^_{}|~]+`).MatchString(s) {
+	if alphanumericRegex.MatchString(s) {
 		// ^[ A-Za-z0-9_@./#&+-]*$/
 		return errors.New(msgAlphanumeric)
 	}