I've wrapped the original `TestParseFileHeader` in order to use the same functionality on the benchmark.
To run the benchmarks I've used the following commands, on `master`:
```
go test -bench=BenchmarkParseFileHeader -count=5 > old
```
On my branch:
```
go test -bench=BenchmarkParseFileHeader -count=5 > new
```
The `benchcmp` output looks like this:
```
% benchcmp old new 
benchmark                      old ns/op     new ns/op     delta
BenchmarkParseFileHeader-8     33891         3308          -90.24%
BenchmarkParseFileHeader-8     33785         3415          -89.89%
BenchmarkParseFileHeader-8     33801         3472          -89.73%
BenchmarkParseFileHeader-8     33830         3494          -89.67%
BenchmarkParseFileHeader-8     33794         3481          -89.70%

benchmark                      old allocs     new allocs     delta
BenchmarkParseFileHeader-8     118            22             -81.36%
BenchmarkParseFileHeader-8     118            22             -81.36%
BenchmarkParseFileHeader-8     118            22             -81.36%
BenchmarkParseFileHeader-8     118            22             -81.36%
BenchmarkParseFileHeader-8     118            22             -81.36%

benchmark                      old bytes     new bytes     delta
BenchmarkParseFileHeader-8     162640        400           -99.75%
BenchmarkParseFileHeader-8     162640        400           -99.75%
BenchmarkParseFileHeader-8     162640        400           -99.75%
BenchmarkParseFileHeader-8     162640        400           -99.75%
BenchmarkParseFileHeader-8     162640        400           -99.75%
```
Cheers